some fixes. moved keyboard names to file. Started adding keyboards for student lists. small refactoring

+++ queue_bot/gdrive_saver.py
@@ -31,8 +31,11 @@ class DriveSaver:
 
         # must create main folder first
         self.init_drive_folder(DriveFolder.HelperBotData)
+
+        # than create other folders
         self.init_drive_folder(DriveFolder.Log)
         self.init_drive_folder(DriveFolder.SubjectChoices)
+        self.init_drive_folder(DriveFolder.Queues)
 
     def init_credentials(self):
         # try read service account file
@@ -42,6 +45,18 @@ class DriveSaver:
             return service_account.Credentials.from_service_account_info(json.loads(os.environ.get('GOOGLE_CREDS')))
         return None
 
+    def init_service(self):
+        if self._credentials is None:
+            print('credentials empty')
+            return None
+
+        try:
+            return discovery.build('drive', 'v3', credentials=self._credentials)
+        except Exception as err:
+            print(err)
+            print('service not initialized')
+            return None
+
     def init_drive_folder(self, drive_folder):
         folder_id = self.load_folder_id(drive_folder)
         if folder_id is None:
@@ -50,19 +65,18 @@ class DriveSaver:
         return folder_id
 
     def create_drive_folder(self, folder_type):
-        if (FolderType.Data.value / folder_type.value).exists():
-            return self.load_folder_id(folder_type)
 
-        folder_metadata = self.get_folder_metadata(folder_type.name)  # use name of enum as folder name
+        service = self.init_service()
+        if service is None:
+            return None
 
-        if folder_type is not DriveFolder.HelperBotData:  # function requests main folder id
-            main_folder_id = self.load_folder_id(DriveFolder.HelperBotData)
-            if main_folder_id is None:
-                main_folder_id = self.create_drive_folder(DriveFolder.HelperBotData)
-            folder_metadata = self.get_folder_metadata(folder_type.name, main_folder_id)
+        if folder_type is DriveFolder.HelperBotData:
+            folder_metadata = DriveSaver.get_folder_metadata(folder_type.name)
+        else:
+            main_folder_id = self.init_drive_folder(DriveFolder.HelperBotData)
+            folder_metadata = DriveSaver.get_folder_metadata(folder_type.name, main_folder_id)
 
         try:
-            service = discovery.build('drive', 'v3', credentials=self._credentials)
             cloud_folder = service.files().create(body=folder_metadata).execute()
             service.permissions().create(fileId=cloud_folder['id'],
                                          body={'type': 'user', 'role': 'writer',
@@ -75,66 +89,59 @@ class DriveSaver:
 
     def save(self, file_path, folder: DriveFolder):
 
-        if self._credentials is None:
+        service = self.init_service()
+        if service is None:
             return False
-
-        try:
-            service = discovery.build('drive', 'v3', credentials=self._credentials)
-        except Exception as err:
-            print(err)
-            return False
-
         folder_id = self.init_drive_folder(folder)
         file_metadata = DriveSaver.get_file_metadata(file_path.name, folder_id)
+        return DriveSaver.upload_to_drive(file_path, file_metadata, service)
 
+    @staticmethod
+    def upload_to_drive(file_path, file_metadata, service):
         try:
             upload_file = MediaFileUpload(file_path, mimetype='application/octet-stream')
             service.files().create(body=file_metadata, media_body=upload_file, fields='id').execute()
         except Exception as err:
             print(err)
             return False
-
         return True
+
+    @staticmethod
+    def update_on_drive(service, file_path, file_id):
+        try:
+            update_file = MediaFileUpload(file_path, mimetype='application/octet-stream')
+            service.files().update(fileId=file_id, media_body=update_file).execute()
+        except HttpError as err:
+            print(err.error_details)
 
     def update_file_list(self, path_list, parent_folder):
 
-        if self._credentials is None:
+        service = self.init_service()
+        if service is None:
             return False
 
-        try:
-            service = discovery.build('drive', 'v3', credentials=self._credentials)
-            existing_files = service.files().list(fields='files(id,name)').execute()['files']
-        except Exception as err:
-            print(err)
-            return False
+        existing_files = DriveSaver.get_existing_files(service, 'id,name')
 
         # update if files exist
-        names_dict = {p.name: p for p in path_list if p.exists()}
+        paths_dict = {p.name: p for p in path_list if p.exists()}
         for file in existing_files:
-            if file['name'] in names_dict:
-                try:
-                    update_file = MediaFileUpload(names_dict[file['name']], mimetype='application/octet-stream')
-                    service.files().update(fileId=file['name'], media_body=update_file).execute()
-                except HttpError as err:
-                    print(err.error_details)
-                finally:
-                    del names_dict[file['name']]  # we delete from dict no matter what error happened
+            if file['name'] in paths_dict:
+                DriveSaver.update_on_drive(service, paths_dict[file['name']], file['id'])
+                del paths_dict[file['name']]
 
         # download to folder if file does not exists
         parent_folder = self.init_drive_folder(parent_folder)
-        for path in names_dict.values():
+        for path in paths_dict.values():
             file_metadata = DriveSaver.get_file_metadata(path.name, parent_folder)
-            try:
-                upload_file = MediaFileUpload(path, mimetype='application/octet-stream')
-                file = service.files().create(body=file_metadata, media_body=upload_file).execute()
-                # service.permissions().create(fileId=file['id'],
-                #                              body={'type': 'user', 'role': 'writer',
-                #                                    'emailAddress': self.work_email}).execute()
-            except HttpError as err:
-                print(err.error_details)
+            DriveSaver.upload_to_drive(path, file_metadata, service)
+        return True
+
+    @staticmethod
+    def get_existing_files(service, fields):
+        return service.files().list(fields='files({0})'.format(fields)).execute()['files']
 
     # if path_list not specified, all files from folder will be written to 'new_folder'
-    def get_file_list(self, path_list, drive_folder):
+    def get_files_from_drive(self, path_list, drive_folder):
         if len(path_list) == 0:
             return True
 
@@ -154,17 +161,15 @@ class DriveSaver:
         return self.download_drive_files(drive_folder_id, path_list)
 
     def download_drive_files(self, folder_id, path_list):
-        try:
-            service = discovery.build('drive', 'v3', credentials=self._credentials)
-            existing_files = service.files().list(fields='files(id,name,parents)').execute()['files']
-        except Exception as err:
-            print(err)
+
+        service = self.init_service()
+        if service is None:
             return False
 
+        existing_files = DriveSaver.get_existing_files(service, 'id,name,parents')
+
         names_list = {p.name: p for p in path_list}
-
         for file in existing_files:
-
             if not file['name'] in names_list:
                 continue
 
@@ -190,18 +195,16 @@ class DriveSaver:
         return True
 
     def clear_drive_folder(self, folder: DriveFolder, exceptions=None):
-
-        folder_id = self.init_drive_folder(folder)
         if exceptions is None:
             exceptions = []
 
-        try:
-            service = discovery.build('drive', 'v3', credentials=self._credentials)
-            existing_files = service.files().list(fields='files(id,name,parents,mimeType)').execute()['files']
-        except Exception as err:
-            print(err)
+        service = self.init_service()
+        if service is None:
             return False
 
+        existing_files = DriveSaver.get_existing_files(service, 'id,name,parents,mimeType')
+
+        folder_id = self.init_drive_folder(folder)
         for file in existing_files:
             if file['name'] in exceptions:
                 continue
@@ -216,23 +219,27 @@ class DriveSaver:
         return True
 
     def delete_everything_on_disk(self):
-        service = discovery.build('drive', 'v3', credentials=self._credentials)
-        existing_files = service.files().list(fields='files(id)').execute()['files']
+
+        service = self.init_service()
+        if service is None:
+            return False
+
+        existing_files = DriveSaver.get_existing_files(service, 'id,name')
 
         for file in existing_files:
             try:
                 result = service.files().delete(fileId=file['id']).execute()
-                print('deleted {0}'.format(result['name']))
+                print('deleted {0}'.format(file['name']))
             except HttpError as err:
                 print(err.content)
 
     def update_all_permissions(self):
-        try:
-            service = discovery.build('drive', 'v3', credentials=self._credentials)
-            existing_files = service.files().list(fields='files(id)').execute()['files']
-        except Exception as err:
-            print(err)
+
+        service = self.init_service()
+        if service is None:
             return False
+
+        existing_files = DriveSaver.get_existing_files(service, 'id')
 
         for file in existing_files:
             try:
@@ -245,12 +252,11 @@ class DriveSaver:
     def show_folder_files(self, folder: DriveFolder):
         folder_id = self.init_drive_folder(folder)
 
-        try:
-            service = discovery.build('drive', 'v3', credentials=self._credentials)
-            existing_files = service.files().list(fields='files(name,parents,id)').execute()['files']
-        except Exception as err:
-            print(err)
+        service = self.init_service()
+        if service is None:
             return False
+
+        existing_files = DriveSaver.get_existing_files(service, 'name,parents,id')
 
         c = 1  # counter of files
         for file in existing_files:
@@ -283,16 +289,17 @@ class DriveSaver:
                 with file.open('r', encoding='utf-8') as fr:
                     cloud_id = fr.read()
             except Exception:
-                print('failed to load id from \"{}\"'.format(file))
+                print('failed to load id from \"{0}\"'.format(file))
 
         return cloud_id
 
     @staticmethod
     def get_file_metadata(name, folder_id):
-        return {
-            'name': name,
-            'parents': [folder_id]
-        }
+        return \
+            {
+                'name': name,
+                'parents': [folder_id]
+            }
 
     @staticmethod
     def get_folder_metadata(name, parent_folder=None):
@@ -313,7 +320,7 @@ class DriveSaver:
 
 if __name__ == '__main__':
     os.chdir(r'D:\coding\Python_codes\Queue_Bot')
-    DriveSaver().delete_everything_on_disk()
+    DriveSaver().show_folder_files(DriveFolder.HelperBotData)
     if len(sys.argv) == 2:
         if sys.argv[1] == 'clear':
             DriveSaver().delete_everything_on_disk()
